+++ unifi/provider.go
@@ -294,6 +294,7 @@ func (p *unifiProvider) Resources(ctx context.Context) []func() resource.Resourc
 		NewFirewallRuleResource,
 		NewFirewallZoneResource,
 		NewNetworkResource,
+		NewOSPFRouterResource,
 		NewPortForwardResource,
 		NewPortProfileFrameworkResource,
 		NewPowerSupervisorResource,